+++ binder.go
@@ -1,8 +1,6 @@
-package htmx
+package xun
 
 import (
-	// "encoding/json"
-
 	"net/http"
 
 	"github.com/go-playground/form/v4"